latency/requirements.txt
ripe.atlas.cousteau==1.4.1
maxminddb==1.5.2
